V-02 / F-02 / F-03 updated
- ApplicationID перенесен в файл .env
- добавлены сервисные файлы для вызова API
- отрисовка табов (без логики переключения) реализована через отдельный компонент v-tabs.
