+++ .gitignore
@@ -22,3 +22,4 @@ pnpm-debug.log*
 *.sln
 *.sw?
 package-lock.json
+.env
